Made requested changes to tests. Changed test names, deleted unnecessary lines, and reorganized methods.
